+++ src/core/CameraController.cpp
@@ -1,4 +1,5 @@
 #include "CameraController.h"
+
 #include <raymath.h>
 #include <rcamera.h>