+++ src/objects/Part.h
@@ -17,9 +17,13 @@ struct Part : Instance {
 	Vector3 Position{0, 0.5, 0};
 	Vector3 Rotation{0, 0, 0};
 	Vector3 Size{2, 1, 4};
+	Vector3 Velocity{0, 0, 0};
 	Color color{163, 162, 165, 255};
 	// where is transparency property?
 	// its baked into the color property since its a raylib color
+	
+	std::string Shape = "Block";
+	bool Anchored = false;
 
 	const char* ClassName() const override {
         return "Part";
@@ -41,6 +45,11 @@ struct Part : Instance {
 			return true;
 		}
 
+		if (std::strcmp(key, "Velocity") == 0) {
+			PushVector3(L, Velocity.x, Velocity.y, Velocity.z);
+			return true;
+		}
+
 		if (std::strcmp(key, "Color") == 0) {
 			PushColor3(L, (float)color.r/255, (float)color.g/255, (float)color.b/255);
 			return true;
@@ -50,29 +59,38 @@ struct Part : Instance {
 			lua_pushnumber(L, 1.f - (float)color.a/255);
 			return true;
 		}
+
+		if (std::strcmp(key, "Shape") == 0) {
+			lua_pushstring(L, Shape.data());
+			return true;
+		}
 		
+		if (std::strcmp(key, "Anchored") == 0) {
+			lua_pushboolean(L, Anchored);
+			return true;
+		}
+
 		return Instance::LuaGet(L, key);
 	}
 	
 	bool LuaSet(lua_State* L, const char* key, int valueIndex) override {
 		if (std::strcmp(key, "Position") == 0) {
-			LuaVector3* vec = CheckVector3(L, valueIndex);
-			Position = {vec->x, vec->y, vec->z};
-
+			Position = RaylibVector3FromLuaVector3(*CheckVector3(L, valueIndex));
 			return true;
 		}
 
 		if (std::strcmp(key, "Rotation") == 0) {
-			LuaVector3* vec = CheckVector3(L, valueIndex);
-			Rotation = {vec->x, vec->y, vec->z};
-			
+			Rotation = RaylibVector3FromLuaVector3(*CheckVector3(L, valueIndex));
 			return true;
 		}
 
 		if (std::strcmp(key, "Size") == 0) {
-			LuaVector3* vec = CheckVector3(L, valueIndex);
-			Size = {vec->x, vec->y, vec->z};
-			
+			Size = RaylibVector3FromLuaVector3(*CheckVector3(L, valueIndex));
+			return true;
+		}
+
+		if (std::strcmp(key, "Velocity") == 0) {
+			Velocity = RaylibVector3FromLuaVector3(*CheckVector3(L, valueIndex));
 			return true;
 		}
 
@@ -89,6 +107,16 @@ struct Part : Instance {
 		if (std::strcmp(key, "Transparency") == 0) {
 			float transparency = static_cast<float>(luaL_checknumber(L, valueIndex));
 			color.a = (unsigned char)((1.f - transparency)*255);
+			return true;
+		}
+
+		if (std::strcmp(key, "Shape") == 0) {
+			Shape = luaL_checkstring(L, valueIndex);
+			return true;
+		}
+
+		if (std::strcmp(key, "Anchored") == 0) {
+			Anchored = luaL_checkboolean(L, valueIndex);
 			return true;
 		}